split index controll into componets
